Move Filter Position/Visibility to Traits
