+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
+## UNRELEASED
+
 ## [v3.1.7] - 2023-12-29
 ### Bug Fixes
 - Fix published view path by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1615
