+++ src/Views/Traits/Helpers/FilterHelpers.php
@@ -88,55 +88,9 @@ trait FilterHelpers
         return $this->getCustomFilterPillValues()[$value] ?? null;
     }
 
-    public function isHiddenFromMenus(): bool
-    {
-        return $this->hiddenFromMenus === true;
-    }
-
-    public function isVisibleInMenus(): bool
-    {
-        return $this->hiddenFromMenus === false;
-    }
-
-    public function isHiddenFromPills(): bool
-    {
-        return $this->hiddenFromPills === true;
-    }
-
-    public function isVisibleInPills(): bool
-    {
-        return $this->hiddenFromPills === false;
-    }
-
-    public function isHiddenFromFilterCount(): bool
-    {
-        return $this->hiddenFromFilterCount === true;
-    }
-
-    public function isVisibleInFilterCount(): bool
-    {
-        return $this->hiddenFromFilterCount === false;
-    }
-
     public function isResetByClearButton(): bool
     {
         return $this->resetByClearButton === true;
-    }
-
-    /**
-     * Determines whether this filter instance is in the secondary header/footer
-     */
-    public function hasCustomPosition(): bool
-    {
-        return ! is_null($this->filterPosition);
-    }
-
-    /**
-     * Returns the custom position of the footer (header or footer)
-     */
-    public function getCustomPosition(): string
-    {
-        return $this->filterPosition;
     }
 
     /**
@@ -153,38 +107,6 @@ trait FilterHelpers
     public function getCustomFilterLabel(): string
     {
         return $this->filterCustomLabel ?? '';
-    }
-
-    /**
-     * Get the filter slide down row.
-     */
-    public function getFilterSlidedownRow(): ?int
-    {
-        return $this->filterSlidedownRow;
-    }
-
-    /**
-     * Get whether the filter has a configured slide down row.
-     */
-    public function hasFilterSlidedownRow(): bool
-    {
-        return ! is_null($this->filterSlidedownRow);
-    }
-
-    /**
-     * Get the filter slide down col span.
-     */
-    public function getFilterSlidedownColspan(): ?int
-    {
-        return $this->filterSlidedownColspan;
-    }
-
-    /**
-     * Get whether the filter has a configured slide down colspan.
-     */
-    public function hasFilterSlidedownColspan(): bool
-    {
-        return ! is_null($this->filterSlidedownColspan);
     }
 
     /**
