+++ src/Views/Traits/IsFilter.php
@@ -4,6 +4,7 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits;
 
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\FilterConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasConfig,HasView};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasCustomPosition,HasVisibility};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\FilterHelpers;
 
 trait IsFilter
@@ -11,17 +12,13 @@ trait IsFilter
     use FilterConfiguration,
         FilterHelpers,
         HasConfig,
+        HasCustomPosition,
+        HasVisibility,
         HasView;
 
     protected string $name;
 
     protected string $key;
-
-    protected bool $hiddenFromMenus = false;
-
-    protected bool $hiddenFromPills = false;
-
-    protected bool $hiddenFromFilterCount = false;
 
     protected bool $resetByClearButton = true;
 
@@ -31,15 +28,9 @@ trait IsFilter
 
     protected array $filterPillValues = [];
 
-    public ?string $filterPosition = null;
-
     protected ?string $filterCustomLabel = null;
 
     protected array $filterLabelAttributes = [];
-
-    protected ?int $filterSlidedownRow = null;
-
-    protected ?int $filterSlidedownColspan = null;
 
     protected ?string $filterCustomPillBlade = null;
